+++ alexber/utils/deploys.py
@@ -4,7 +4,7 @@ This module is usable in your deployment script. See also `fabs` module.
 See here https://medium.com/analytics-vidhya/my-ymlparsers-module-88221edf16a6 for documentation.
 
 This module depends on some 3-rd party dependencies, in order to use it, you should have installed first. To do it
-run `python3 -m pip install alex-ber-utils[yml]`.
+run `python -m pip install alex-ber-utils[yml]`.
 """
 
 from collections import OrderedDict