+++ CHANGELOG.md
@@ -3,7 +3,14 @@ All notable changes to this project will be documented in this file.
 
 \#https://pypi.org/manage/project/alex-ber-utils/releases/
 
+
 ## Unreleased
+
+## [0.11.0a0] 20.06.2024
+TBD
+- Removed leftover from multidispatch dependecy.
+- Rename python3 to python accross the documentation.
+
 ## [0.10.2] 05.05.2024
 - Lose packaging constraint to packaging>=23.2.
 - Fixed README.md to include packaging>=23.2.